102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -1,4 +1,3 @@
-#include "main.h"
 #include <stdio.h>
 /**
  * main - check the code
@@ -7,14 +6,15 @@
  */
 int main(void)
 {
-long long i = 1, j = 2, k, count = 50;
-printf("%lld, %lld, ", i, j);
+unsigned long i = 1, j = 2, k, count = 50;
+
+printf("%lu, %lu, ", i, j);
 for (k = 0; k < count; k++)
 {
 i = i + j;
-printf("%lld, ", i);
+printf("%lu, ", i);
 j = i + j;
-printf("%lld, ", j);
+printf("%lu, ", j);
 }
 return (0);
 }
